feat(FE): 브라우저 캐시 무효화 및 webpack 빌드 환경 분리 (#307)
* feat(FE): 웹팩 설정 최적화 및 청크 분할 추가

* feat(FE): 웹팩 설정 파일 분리 및 환경별 설정 추가

* feat(FE): 웹팩 스크립트 환경 변수 추가

* feat(FE): 웹팩 설정에서 Sentry 플러그인 임포트 위치 수정

* feat(FE): 웹팩 설정에 webpack-merge 추가 및 __dirname 내보내기 수정

* feat(FE): 웹팩 최적화 설정에서 minimize 옵션 추가

* refactor(FE): 웹팩 설정에서 중복 merge 수정
