+++ frontend/webpack.common.js
@@ -1,0 +1,52 @@
+import path from "node:path";
+import { fileURLToPath } from "node:url";
+import Dotenv from "dotenv-webpack";
+import ForkTsCheckerWebpackPlugin from "fork-ts-checker-webpack-plugin";
+import HtmlWebpackPlugin from "html-webpack-plugin";
+
+const __filename = fileURLToPath(import.meta.url);
+export const __dirname = path.dirname(__filename);
+
+export default {
+  entry: "./src/main.tsx",
+  output: {
+    publicPath: "/",
+    path: path.resolve(__dirname, "dist"),
+    clean: true,
+  },
+  resolve: {
+    extensions: [".ts", ".tsx", ".js"],
+    alias: {
+      "@shared": path.resolve(__dirname, "src/shared"),
+      "@assets": path.resolve(__dirname, "src/assets"),
+      "@domains": path.resolve(__dirname, "src/domains"),
+      "@": path.resolve(__dirname, "src"),
+    },
+  },
+
+  module: {
+    rules: [
+      {
+        test: /\.(ts|tsx)$/,
+        exclude: /node_modules/,
+        use: {
+          loader: "babel-loader",
+        },
+      },
+      {
+        test: /\.(png|jpe?g|gif|svg|webp)$/i,
+        type: "asset/resource",
+        generator: {
+          filename: "assets/[name][hash][ext]",
+        },
+      },
+    ],
+  },
+  plugins: [
+    new HtmlWebpackPlugin({
+      template: "./public/index.html",
+    }),
+    new ForkTsCheckerWebpackPlugin(),
+    new Dotenv({ path: ".env.local", systemvars: true }),
+  ],
+};
